Initial event implementation
Add initial implementation of the simulator event loop.

Create Host class which holds information about the system, i.e. the
assignment of requests to cores and whether a core is idle or not.

Create Request class which holds information about individual
requests, i.e. their id and execution time.

Create Request_Generator class which generates requests and triggers
most of the events of the system in the other classes. Currently,
generate the requests with constant execution time and at constant
intervals.

Create Scheduler class which is triggered every time a new request
arrives to the system or a request finishes its execution. While there
is a request in the queue and a free core assign it directly to it,
otherwise yield immediately.

Signed-off-by: Kostis Kaffes <kkaffes@stanford.edu>