Add Shinjuku scheduler and cpu group
Shinjuku shceduler contains CPU group that is shared with the
global queue host configuration. In each CPU group, there are
active and idle cores so that it's constant time lookup for idle
core.

Each CPU controlled by the shinjuku scheduler is called a
notificationcore, which basically runs for a certain time and
notifies the scheduler when done (essentially activating
scheduler and adding it back to idle core)